separate php composer resolution and parsing
